requirements.txt
asgiref==3.2.10
Django==3.1.2
django-cors-headers==3.6.0
django-filter==2.4.0
djangorestframework==3.12.2
flake8==3.8.4
gunicorn==20.0.4
Markdown==3.3.3
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2018.7
sqlparse==0.4.1
